Move application form data types and editor validations to lib-common
