+++ frontend/src/citizen-frontend/applications/editor/verification/AdditionalDetailsSection.tsx
@@ -5,7 +5,7 @@
 import React from 'react'
 import styled from 'styled-components'
 
-import type { ApplicationFormData } from 'lib-common/api-types/application/ApplicationFormData'
+import type { ApplicationFormData } from 'lib-common/application/ApplicationFormData'
 import ListGrid from 'lib-components/layout/ListGrid'
 import { H2, Label } from 'lib-components/typography'
 import { Gap } from 'lib-components/white-space'
