+++ frontend/src/citizen-frontend/applications/read-view/ApplicationReadViewContents.tsx
@@ -4,7 +4,7 @@
 
 import React, { useContext } from 'react'
 
-import type { ApplicationFormData } from 'lib-common/api-types/application/ApplicationFormData'
+import type { ApplicationFormData } from 'lib-common/application/ApplicationFormData'
 import type { ApplicationDetails as ApplicationDetailsGen } from 'lib-common/generated/api-types/application'
 import HorizontalLine from 'lib-components/atoms/HorizontalLine'
 import Main from 'lib-components/atoms/Main'
