+++ frontend/src/e2e-test/utils/application-forms.ts
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
-import type { ApplicationFormData } from 'lib-common/api-types/application/ApplicationFormData'
+import type { ApplicationFormData } from 'lib-common/application/ApplicationFormData'
 import type {
   ApplicationDetails,
   OtherGuardianAgreementStatus,
